Bob: Write partition entries
Write out the partition entries. Additionally make sure to seek the
File cursor to the correct spot when completing the write for the
partition table header. This is also a nice milestone because gdsik
doesn't find anything in particular wrong with a simple 1 partiton
disk image generated with Bob! I'm sure there's tons of bugs, but
it "sorta works" :)